+++ examples/print.rs
@@ -1,7 +1,7 @@
-use termal::{printcln, reset_terminal};
+use termal::{codes, printcln, reset_terminal};
 
 fn main() {
-    printcln!("\x1b]4;220;?\x07");
+    printcln!("{}", codes::set_selection([], b"hello there"));
     _ = std::io::stdin().read_line(&mut String::new());
     reset_terminal();
 }