+++ termal_core/src/codes.rs
@@ -20,12 +20,18 @@
 //!   to terminal to do what they say they do. This is the majority of the
 //!   codes.
 
+use std::fmt::Display;
+
+use base64::Engine;
 use place_macro::place;
 
 /// Creates the given sequence, this is used internally, you should use
 /// the macro [`csi`]
 #[macro_export]
 macro_rules! seq {
+    ($sq:literal, $i:literal) => {
+        concat!($sq, $i)
+    };
     ($sq:literal, $i:literal, $f:literal, $($a:literal),*) => {
         concat!($sq, $f $(, ';', $a)*, $i)
     };
@@ -45,19 +51,52 @@ pub const ESC: char = '\x1b';
 pub const CSI: &str = "\x1b[";
 /// Device Control String: Start of DCS sequence
 pub const DCS: &str = "\x1bP";
-/// Operating System Command: Start of OCS sequence
-pub const OCS: &str = "\x1b]";
+/// Operating System Command: Start of OSC sequence
+pub const OSC: &str = "\x1b]";
 /// String terminator. Terminates for example DCS.
 pub const ST: &str = "\x1b\\";
 /// Single shift three
 pub const SS3: &str = "\x1bO";
 
-/// Creates escape sequence, the first literal is the end of the sequence,
-/// the other arguments are the values in the sequence
+/// Creates control escape sequence, the first literal is the end of the
+/// sequence, the other arguments are the values in the sequence
 #[macro_export]
 macro_rules! csi {
-    ($i:literal, $($a:expr),+) => {
-        $crate::seq!("\x1b[", $i, $($a),+)
+    ($i:literal $(,$a:expr)* $(,)?) => {
+        $crate::seq!("\x1b[", $i $(, $a)*)
+    };
+}
+
+/// Creates control escape sequence for graphic mode.
+#[macro_export]
+macro_rules! graphic {
+    ($($a:expr),* $(,)?) => {
+        $crate::csi!('m' $(, $a)*)
+    };
+}
+
+/// Creates operating system command sequence. The arguments are the values in
+/// the sequence.
+#[macro_export]
+macro_rules! osc {
+    ($($a:expr),+) => {
+        $crate::seq!("\x1b]", "\x1b\\", $($a),+)
+    };
+}
+
+/// Enables the given private terminal mode.
+#[macro_export]
+macro_rules! enable {
+    ($a:expr) => {
+        $crate::seq!("\x1b[?", 'h', $a)
+    };
+}
+
+/// Disables the given private terminal mode.
+#[macro_export]
+macro_rules! disable {
+    ($a:expr) => {
+        $crate::seq!("\x1b[?", 'l', $a)
     };
 }
 
@@ -88,27 +127,27 @@ pub const DELETE: char = '\x7f';
 // `.get_string()` method from the trait [`GetString`] to get [`String`] in
 // both cases
 
-macro_rules! csi_macro {
-    ($(
+macro_rules! code_macro {
+    ($code:ident $(
         $name:ident
         $(, $($nam:ident)? $($lit:literal)?)+ ;
-        $i:literal $(?$doc:literal)?),+ $(,)?
+        $($i:literal)? $(?$doc:literal)?),+ $(,)?
     ) => {
         place! {$(
             $(#[doc = __repnl__($doc, " ")])?
             #[macro_export]
             macro_rules! $name {
                 (__start__($($(__s__ $nam:expr,)?)+) __s__ (,)?) => {
-                    __s__ crate::csi!($i, $($(__s__ $nam)? $($lit)?),+)
+                    __s__ crate::$code!($($i,)? $($(__s__ $nam)? $($lit)?),+)
                 }
             }
             pub use $name;
         )+}
     };
-    (!= $ex:literal => $(
+    ($code:ident != $ex:literal => $(
         $name:ident,
         $nam:ident;
-        $i:literal $(?$doc:literal)?),+ $(,)?
+        $($i:literal)? $(?$doc:literal)?),+ $(,)?
     ) => {
         place! {$(
             $(#[doc = __repnl__($doc, " ")])?
@@ -119,7 +158,7 @@ macro_rules! csi_macro {
                     if v == $ex {
                         "".into()
                     } else {
-                        __s__ crate::csi!($i, v)
+                        __s__ crate::$code!($($i,)? v)
                     }
                 }}
             }
@@ -138,7 +177,9 @@ macro_rules! move_to {
 
 pub use move_to;
 
-csi_macro!( != 0 =>
+use crate::Rgb;
+
+code_macro!(csi != 0 =>
     move_up, n; 'A' ? "Moves cursor up by N positions",
     move_down, n; 'B' ? "Moves cursor down by N positions",
     move_right, n; 'C' ? "Moves cursor right by N positions",
@@ -152,9 +193,10 @@ csi_macro!( != 0 =>
     delete_columns, n; "'~" ? "Delete n columns, moving them from the right",
     set_down, n; 'E' ? "Moves cursor to the start of line N lines down",
     set_up, n; 'F' ? "Moves cursor to the start of line N lines up",
+    repeat_char, n; 'b' ? "Repeat the previous char n times."
 );
 
-csi_macro!(
+code_macro!(csi
     column, n; 'G' ? "Moves cursor to the given column",
 );
 
@@ -169,166 +211,163 @@ pub const CUR_LOAD: &str = "\x1b8";
 // Erase codes
 
 /// Erases from the cursor to the end of the screen
-pub const ERASE_TO_END: &str = "\x1b[J";
+pub const ERASE_TO_END: &str = csi!('J');
 /// Erases from the start of the screen to the cursor
-pub const ERASE_FROM_START: &str = "\x1b[1J";
+pub const ERASE_FROM_START: &str = csi!('J', 1);
 /// Erases the entire screen
-pub const ERASE_SCREEN: &str = "\x1b[2J";
+pub const ERASE_SCREEN: &str = csi!('J', 2);
 /// Erases the whole screen and the scrollback buffer
-pub const ERASE_ALL: &str = "\x1b[3J";
+pub const ERASE_ALL: &str = csi!('J', 3);
 /// Erases from cursor to the end of the line
-pub const ERASE_TO_LN_END: &str = "\x1b[K";
+pub const ERASE_TO_LN_END: &str = csi!('K');
 /// Erases from the start of the line to the cursor
-pub const ERASE_FROM_LN_START: &str = "\x1b[1K";
+pub const ERASE_FROM_LN_START: &str = csi!('K', 1);
 /// Erases the entire line
-pub const ERASE_LINE: &str = "\x1b[2K";
+pub const ERASE_LINE: &str = csi!('K', 2);
 
 // Text modes
 
 /// Resets all the text modes (colors and styles)
-pub const RESET: &str = "\x1b[0m";
+pub const RESET: &str = graphic!(0);
 
 /// Set bold text mode (on some terminals may be just brighter color)
-pub const BOLD: &str = "\x1b[1m";
+pub const BOLD: &str = graphic!(1);
 /// Set dim/faint text mode
-pub const FAINT: &str = "\x1b[2m";
+pub const FAINT: &str = graphic!(2);
 /// Set italic mode
-pub const ITALIC: &str = "\x1b[3m";
+pub const ITALIC: &str = graphic!(3);
 /// Set underline mode
-pub const UNDERLINE: &str = "\x1b[4m";
+pub const UNDERLINE: &str = graphic!(4);
 /// Set blinking mode
-pub const BLINKING: &str = "\x1b[5m";
+pub const BLINKING: &str = graphic!(5);
 /// Set inverse mode (inverse foreground and background)
-pub const INVERSE: &str = "\x1b[7m";
+pub const INVERSE: &str = graphic!(7);
 /// Set invisible mode (foreground is same as background)
-pub const INVISIBLE: &str = "\x1b[8m";
+pub const INVISIBLE: &str = graphic!(8);
 /// Set striketrough mode
-pub const STRIKETROUGH: &str = "\x1b[9m";
+pub const STRIKETROUGH: &str = graphic!(9);
 /// Set double underline mode
-pub const DOUBLE_UNDERLINE: &str = "\x1b[21";
+pub const DOUBLE_UNDERLINE: &str = graphic!(21);
 /// Set ouverline mode
-pub const OVERLINE: &str = "\x1b[53";
+pub const OVERLINE: &str = graphic!(53);
 
 /// Reset [`BOLD`] and [`FAINT`] mode
-pub const RESET_BOLD: &str = "\x1b[22m";
+pub const RESET_BOLD: &str = graphic!(22);
 /// Reset [`ITALIC`] mode
-pub const RESET_ITALIC: &str = "\x1b[23m";
+pub const RESET_ITALIC: &str = graphic!(23);
 /// Reset [`UNDERLINE`] and [`DOUBLE_UNDERLINE`] mode
-pub const RESET_UNDERLINE: &str = "\x1b[24m";
+pub const RESET_UNDERLINE: &str = graphic!(24);
 /// Reset [`BLINKING`] mode
-pub const RESET_BLINKING: &str = "\x1b[25m";
+pub const RESET_BLINKING: &str = graphic!(25);
 /// Reset [`INVERSE`] mode
-pub const RESET_INVERSE: &str = "\x1b[27m";
+pub const RESET_INVERSE: &str = graphic!(27);
 /// Reset [`INVISIBLE`] mode
-pub const RESET_INVISIBLE: &str = "\x1b[28m";
+pub const RESET_INVISIBLE: &str = graphic!(28);
 /// Reset [`STRIKETROUGH`] mode
-pub const RESET_STRIKETROUGH: &str = "\x1b[29m";
+pub const RESET_STRIKETROUGH: &str = graphic!(29);
 /// Reset [`OVERLINE`] mode.
-pub const RESET_OVERLINE: &str = "\x1b[55m";
+pub const RESET_OVERLINE: &str = graphic!(55);
 
 /// Set the foreground color to black (dark black)
-pub const BLACK_FG: &str = "\x1b[30m";
+pub const BLACK_FG: &str = graphic!(30);
 /// Set the foreground color to white (bright white)
-pub const WHITE_FG: &str = "\x1b[97m";
+pub const WHITE_FG: &str = graphic!(97);
 /// Set the foreground color to gray (bright black)
-pub const GRAY_FG: &str = "\x1b[90m";
+pub const GRAY_FG: &str = graphic!(90);
 /// Set to foreground color to bright gray (dark white)
-pub const GRAY_BRIGHT_FG: &str = "\x1b[37m";
+pub const GRAY_BRIGHT_FG: &str = graphic!(37);
 
 /// Set the foreground color to red (bright red)
-pub const RED_FG: &str = "\x1b[91m";
+pub const RED_FG: &str = graphic!(91);
 /// Set the foreground color to green (bright green)
-pub const GREEN_FG: &str = "\x1b[92m";
+pub const GREEN_FG: &str = graphic!(92);
 /// Set the foreground color to yellow (bright yellow)
-pub const YELLOW_FG: &str = "\x1b[93m";
+pub const YELLOW_FG: &str = graphic!(93);
 /// Set the foreground color to blue (bright blue)
-pub const BLUE_FG: &str = "\x1b[94m";
+pub const BLUE_FG: &str = graphic!(94);
 /// Set the foreground color to magenta (bright magenta)
-pub const MAGENTA_FG: &str = "\x1b[95m";
+pub const MAGENTA_FG: &str = graphic!(95);
 /// Set the foreground color to cyan (bright cyan)
-pub const CYAN_FG: &str = "\x1b[96m";
+pub const CYAN_FG: &str = graphic!(96);
 
 /// Set the foreground color to dark red
-pub const RED_DARK_FG: &str = "\x1b[31m";
+pub const RED_DARK_FG: &str = graphic!(31);
 /// Set the foreground color to dark green
-pub const GREEN_DARK_FG: &str = "\x1b[32m";
+pub const GREEN_DARK_FG: &str = graphic!(32);
 /// Set the foreground color to dark yellow
-pub const YELLOW_DARK_FG: &str = "\x1b[33m";
+pub const YELLOW_DARK_FG: &str = graphic!(33);
 /// Set the foreground color to dark blue
-pub const BLUE_DARK_FG: &str = "\x1b[34m";
+pub const BLUE_DARK_FG: &str = graphic!(34);
 /// Set the foreground color to dark magenta
-pub const MAGENTA_DARK_FG: &str = "\x1b[35m";
+pub const MAGENTA_DARK_FG: &str = graphic!(35);
 /// Set the foreground color to dark cyan
-pub const CYAN_DARK_FG: &str = "\x1b[36m";
+pub const CYAN_DARK_FG: &str = graphic!(36);
 
 /// Reset the foreground color
-pub const RESET_FG: &str = "\x1b[39m";
+pub const RESET_FG: &str = graphic!(39);
 
 /// Set the background color to black (dark black)
-pub const BLACK_BG: &str = "\x1b[40m";
+pub const BLACK_BG: &str = graphic!(40);
 /// Set the background color to white (bright white)
-pub const WHITE_BG: &str = "\x1b[107m";
+pub const WHITE_BG: &str = graphic!(107);
 /// Set the background color to gray (bright black)
-pub const GRAY_BG: &str = "\x1b[100m";
+pub const GRAY_BG: &str = graphic!(100);
 /// Set to background color to bright gray (dark white)
-pub const GRAY_BRIGHT_BG: &str = "\x1b[47m";
+pub const GRAY_BRIGHT_BG: &str = graphic!(47);
 
 /// Set the background color to red (bright red)
-pub const RED_BG: &str = "\x1b[101m";
+pub const RED_BG: &str = graphic!(101);
 /// Set the background color to green (bright green)
-pub const GREEN_BG: &str = "\x1b[102m";
+pub const GREEN_BG: &str = graphic!(102);
 /// Set the background color to yellow (bright yellow)
-pub const YELLOW_BG: &str = "\x1b[103m";
+pub const YELLOW_BG: &str = graphic!(103);
 /// Set the background color to blue (bright blue)
-pub const BLUE_BG: &str = "\x1b[104m";
+pub const BLUE_BG: &str = graphic!(104);
 /// Set the background color to magenta (bright magenta)
-pub const MAGENTA_BG: &str = "\x1b[105m";
+pub const MAGENTA_BG: &str = graphic!(105);
 /// Set the background color to cyan (bright cyan)
-pub const CYAN_BG: &str = "\x1b[106m";
+pub const CYAN_BG: &str = graphic!(106);
 
 /// Set the background color to dark red
-pub const RED_DARK_BG: &str = "\x1b[41m";
+pub const RED_DARK_BG: &str = graphic!(41);
 /// Set the background color to dark green
-pub const GREEN_DARK_BG: &str = "\x1b[42m";
+pub const GREEN_DARK_BG: &str = graphic!(42);
 /// Set the background color to dark yellow
-pub const YELLOW_DARK_BG: &str = "\x1b[43m";
+pub const YELLOW_DARK_BG: &str = graphic!(43);
 /// Set the background color to dark blue
-pub const BLUE_DARK_BG: &str = "\x1b[44m";
+pub const BLUE_DARK_BG: &str = graphic!(44);
 /// Set the background color to dark magenta
-pub const MAGENTA_DARK_BG: &str = "\x1b[45m";
+pub const MAGENTA_DARK_BG: &str = graphic!(45);
 /// Set the background color to dark cyan
-pub const CYAN_DARK_BG: &str = "\x1b[46m";
+pub const CYAN_DARK_BG: &str = graphic!(46);
 
 /// Reset the background color
-pub const RESET_BG: &str = "\x1b[49m";
+pub const RESET_BG: &str = graphic!(49);
 
-csi_macro! {
-    fg256, 38, 5, c; 'm'
+code_macro! { graphic
+    fg256, 38, 5, c;
         ? "creates a foreground color, color is value in range 0..256",
 
-    bg256, 48, 5, c; 'm'
+    bg256, 48, 5, c;
         ? "creates a background color, color is value in range 0..256",
 
-    underline256, 58, 5, c; 'm'
+    underline256, 58, 5, c;
         ? "Set underline color as 256 color.",
 
-    fg, 38, 2, r, g, b; 'm'
+    fg, 38, 2, r, g, b;
         ? "creates a true rgb foreground color. R, G and B must be values in
            range 0..256",
 
-    bg, 48, 2, r, g, b; 'm'
+    bg, 48, 2, r, g, b;
         ? "creates a true rgb background color. R, G and B must be values in
            range 0..256",
 
-    underline_rgb, 58, 2, r, g, b; 'm'
+    underline_rgb, 58, 2, r, g, b;
         ? "Set underline color as rgb.",
-
-    repeat_char, n; 'b'
-        ? "Repeat the previous char n times."
 }
 
 /// Reset the underline color.
-pub const RESET_UNDERLINE_COLOR: &str = "\x1b[59m";
+pub const RESET_UNDERLINE_COLOR: &str = graphic!(59);
 
 // Line modes
 /// Makes this line characters twice as large overlapping with the line below.
@@ -348,37 +387,37 @@ pub const ENABLE_LINE_WRAP: &str = "\x1b[=7h";
 pub const DISABLE_LINE_WRAP: &str = "\x1b[=7l";
 
 /// Enables reverse color for the whole terminal display.
-pub const ENABLE_REVERSE_COLOR: &str = "\x1b[?5h";
+pub const ENABLE_REVERSE_COLOR: &str = enable!(5);
 /// Disables reverse color for the whole terminal display. (This actually often
 /// doesn't work)
-pub const DISABLE_REVERSE_COLOR: &str = "\x1b[?5l";
+pub const DISABLE_REVERSE_COLOR: &str = disable!(5);
 
 // Private modes
 
 /// Makes the cursor invisible
-pub const HIDE_CURSOR: &str = "\x1b[?25l";
+pub const HIDE_CURSOR: &str = disable!(25);
 /// Makes the cursor visible
-pub const SHOW_CURSOR: &str = "\x1b[?25h";
+pub const SHOW_CURSOR: &str = enable!(25);
 /// Saves the visible part of the screen buffer
-pub const SAVE_SCREEN: &str = "\x1b[?47l";
+pub const SAVE_SCREEN: &str = disable!(47);
 /// Loads the last saved screen
-pub const LOAD_SCREEN: &str = "\x1b[?47h";
+pub const LOAD_SCREEN: &str = enable!(47);
 /// Enables alternative buffer
-pub const ENABLE_ALTERNATIVE_BUFFER: &str = "\x1b[?1049h";
+pub const ENABLE_ALTERNATIVE_BUFFER: &str = enable!(1049);
 /// Disables the laternative buffer
-pub const DISABLE_ALTERNATIVE_BUFFER: &str = "\x1b[?1049l";
+pub const DISABLE_ALTERNATIVE_BUFFER: &str = disable!(1049);
 
 // Other
 /// Full terminal reset. Clear the screen, buffer, reset all modes, ...
 pub const FULL_RESET: &str = "\x1bc";
 
 /// Request the device attributes.
-pub const REQUEST_DEVICE_ATTRIBUTES: &str = "\x1b[c";
+pub const REQUEST_DEVICE_ATTRIBUTES: &str = csi!('c');
 /// Request the device status.
-pub const REQUEST_STATUS_REPORT: &str = "\x1b[5n";
+pub const REQUEST_STATUS_REPORT: &str = csi!('n', 5);
 /// Request the current cursor position. In some terminals, the report may be
 /// ambigous with F3 key press with modifiers.
-pub const REQUEST_CURSOR_POSITION: &str = "\x1b[6n";
+pub const REQUEST_CURSOR_POSITION: &str = csi!('n', 6);
 /// Request the current cursor position. Difference from
 /// [`REQUEST_CURSOR_POSITION`] is that the response is not ambigous, but it is
 /// not supported by some terminals that support [`REQUEST_CURSOR_POSITION`].
@@ -386,77 +425,77 @@ pub const REQUEST_CURSOR_POSITION2: &str = "\x1b[?6n";
 /// Requests the terminal name and version.
 pub const REQUEST_TERMINAL_NAME: &str = "\x1b[>0q";
 /// Request the text area size of terminal in pixels.
-pub const REQUEST_TEXT_AREA_SIZE_PX: &str = "\x1b[14t";
+pub const REQUEST_TEXT_AREA_SIZE_PX: &str = csi!('t', 14);
 /// Request size of single character on creen in pixels.
-pub const REQUEST_CHAR_SIZE: &str = "\x1b[16t";
+pub const REQUEST_CHAR_SIZE: &str = csi!('t', 16);
 /// Request size of the text area in characters.
-pub const REQUEST_TEXT_AREA_SIZE: &str = "\x1b[18t";
+pub const REQUEST_TEXT_AREA_SIZE: &str = csi!('t', 18);
 /// Request the number of sixel color registers.
-pub const REQUEST_SIXEL_COLORS: &str = "\x1b[1;1;1S";
+pub const REQUEST_SIXEL_COLORS: &str = csi!('S', 1, 1, 1);
 
 /// Enables mouse tracking for X and Y coordinate on press.
-pub const ENABLE_MOUSE_XY_TRACKING: &str = "\x1b[?9h";
+pub const ENABLE_MOUSE_XY_TRACKING: &str = enable!(9);
 /// Disables mouse tracking for X and Y coordinate on press.
-pub const DISABLE_MOUSE_XY_TRACKING: &str = "\x1b[?9l";
+pub const DISABLE_MOUSE_XY_TRACKING: &str = disable!(9);
 /// Enables mouse tracking for X and Y coordinate on press and release.
-pub const ENABLE_MOUSE_XY_PR_TRACKING: &str = "\x1b[?1000h";
+pub const ENABLE_MOUSE_XY_PR_TRACKING: &str = enable!(1000);
 /// Disables mouse tracking for X and Y coordinate on press and release.
-pub const DISABLE_MOUSE_XY_PR_TRACKING: &str = "\x1b[?1000l";
+pub const DISABLE_MOUSE_XY_PR_TRACKING: &str = disable!(1000);
 /// Enables mouse tracking for X and Y coordinate on press, release and drag.
-pub const ENABLE_MOUSE_XY_DRAG_TRACKING: &str = "\x1b[?1002h";
+pub const ENABLE_MOUSE_XY_DRAG_TRACKING: &str = enable!(1002);
 /// Disables mouse tracking for X and Y coordinate on press, release and drag.
-pub const DISABLE_MOUSE_XY_DRAG_TRACKING: &str = "\x1b[?1002l";
+pub const DISABLE_MOUSE_XY_DRAG_TRACKING: &str = disable!(1002);
 /// Enables mouse tracking for X and Y coordinate on press, release, drag and
 /// move.
-pub const ENABLE_MOUSE_XY_ALL_TRACKING: &str = "\x1b[?1002h";
+pub const ENABLE_MOUSE_XY_ALL_TRACKING: &str = enable!(1003);
 /// Disables mouse tracking for X and Y coordinate on press, release, drag and
 /// move.
-pub const DISABLE_MOUSE_XY_ALL_TRACKING: &str = "\x1b[?1002l";
+pub const DISABLE_MOUSE_XY_ALL_TRACKING: &str = disable!(1003);
 /// Enables sending event on focus gain.
-pub const ENABLE_FOCUS_EVENT: &str = "\x1b[?1004h";
+pub const ENABLE_FOCUS_EVENT: &str = enable!(1004);
 /// Disables sending event on focus gain.
-pub const DISABLE_FOCUS_EVENT: &str = "\x1b[?1004l";
+pub const DISABLE_FOCUS_EVENT: &str = disable!(1004);
 /// Enables extension to send mouse inputs in format extended to utf8 two byte
 /// characters.
-pub const ENABLE_MOUSE_XY_UTF8_EXT: &str = "\x1b[?1005h";
+pub const ENABLE_MOUSE_XY_UTF8_EXT: &str = enable!(1005);
 /// Disables extension to send mouse inputs in format extended to utf8 two byte
 /// characters.
-pub const DISABLE_MOUSE_XY_UTF8_EXT: &str = "\x1b[?1005l";
+pub const DISABLE_MOUSE_XY_UTF8_EXT: &str = disable!(1005);
 /// Enables extension to send mouse inputs in different format as position in
 /// characters.
-pub const ENABLE_MOUSE_XY_EXT: &str = "\x1b[?1006h";
+pub const ENABLE_MOUSE_XY_EXT: &str = enable!(1006);
 /// Disables extension to send mouse inputs in different format as position in
 /// characters.
-pub const DISABLE_MOUSE_XY_EXT: &str = "\x1b[?1006l";
+pub const DISABLE_MOUSE_XY_EXT: &str = disable!(1006);
 /// Enables URXVT mouse extension. Not recommended, rather use
 /// [`ENABLE_MOUSE_XY_EXT`].
-pub const ENABLE_MOUSE_XY_URXVT_EXT: &str = "\x1b[?1015h";
+pub const ENABLE_MOUSE_XY_URXVT_EXT: &str = enable!(1015);
 /// Disables URXVT mouse extension.
-pub const DISABLE_MOUSE_XY_URXVT_EXT: &str = "\x1b[?1015l";
+pub const DISABLE_MOUSE_XY_URXVT_EXT: &str = disable!(1015);
 /// Enables extension to send mouse inputs in different format as position in
 /// pixels.
-pub const ENABLE_MOUSE_XY_PIX_EXT: &str = "\x1b[?1016h";
+pub const ENABLE_MOUSE_XY_PIX_EXT: &str = enable!(1016);
 /// Disables extension to send mouse inputs in different format as position in
 /// pixels.
-pub const DISABLE_MOUSE_XY_PIX_EXT: &str = "\x1b[?1016l";
+pub const DISABLE_MOUSE_XY_PIX_EXT: &str = disable!(1016);
 
-csi_macro! {
+code_macro! { csi
     scroll_region, t, b; 'r'
         ? "Set the scroll region in the terminal. Also moves the cursor to the
            top left."
 }
 
 /// Reset the scroll region
-pub const RESET_SCROLL_REGION: &str = "\x1b[0;0r";
+pub const RESET_SCROLL_REGION: &str = scroll_region!(0, 0);
 /// Don't limit the printing area.
-pub const DONT_LIMIT_PRINT_TO_SCROLL_REGION: &str = "\x1b[19h";
+pub const DONT_LIMIT_PRINT_TO_SCROLL_REGION: &str = enable!(19);
 /// Limit printing area only to scroll region.
-pub const LIMIT_PRINT_TO_SCROLL_REGION: &str = "\x1b[19l";
+pub const LIMIT_PRINT_TO_SCROLL_REGION: &str = disable!(19);
 
 /// Enables bracketed paste mode. In this mode, pasted text is treated
 /// verbatim.
-pub const ENABLE_BRACKETED_PASTE_MODE: &str = "\x1b[?2004h";
-pub const DISABLE_BRACKETED_PASTE_MODE: &str = "\x1b[?2004l";
+pub const ENABLE_BRACKETED_PASTE_MODE: &str = enable!(2004);
+pub const DISABLE_BRACKETED_PASTE_MODE: &str = disable!(2004);
 
 #[derive(Clone, Debug, Copy, Eq, PartialEq)]
 pub enum CursorStyle {
@@ -477,15 +516,141 @@ pub enum CursorStyle {
 
 pub fn set_cursor(style: CursorStyle) -> &'static str {
     match style {
-        CursorStyle::Block(Some(true)) => "\x1b[0 q",
-        CursorStyle::Block(None) => "\x1b[1 q",
-        CursorStyle::Block(Some(false)) => "\x1b[2 q",
-        CursorStyle::Underline(true) => "\x1b[3 q",
-        CursorStyle::Underline(false) => "\x1b[4 q",
-        CursorStyle::Bar(true) => "\x1b[5 q",
-        CursorStyle::Bar(false) => "\x1b[6 q",
+        CursorStyle::Block(Some(true)) => csi!(" q", 0),
+        CursorStyle::Block(None) => csi!(" q", 1),
+        CursorStyle::Block(Some(false)) => csi!(" q", 2),
+        CursorStyle::Underline(true) => csi!(" q", 3),
+        CursorStyle::Underline(false) => csi!(" q", 4),
+        CursorStyle::Bar(true) => csi!(" q", 5),
+        CursorStyle::Bar(false) => csi!(" q", 6),
     }
 }
+
+// OSC sequences
+
+code_macro! {osc
+    request_color_code, 4, code, "?";
+        ? "Requests the current color assigned to the given color code.",
+
+    reset_color_code, 104, code;
+        ? "Resets the color definition for the given color code.",
+}
+
+/// Defines color for the given color code.
+pub fn define_color_code<T>(code: u8, color: impl Into<Rgb<T>>) -> String
+where
+    Rgb<T>: Display,
+{
+    osc!(4, code, color.into())
+}
+
+/// Sets the default foreground color
+pub fn set_default_fg_color<T>(color: impl Into<Rgb<T>>) -> String
+where
+    Rgb<T>: Display,
+{
+    osc!(10, color.into())
+}
+
+/// Sets the default foreground color
+pub fn set_default_bg_color<T>(color: impl Into<Rgb<T>>) -> String
+where
+    Rgb<T>: Display,
+{
+    osc!(11, color.into())
+}
+
+/// Sets the color of the cursor.
+pub fn set_cursor_color<T>(color: impl Into<Rgb<T>>) -> String
+where
+    Rgb<T>: Display,
+{
+    osc!(12, color.into())
+}
+
+/// Resets all the color codes to their default colors.
+pub const RESET_ALL_COLOR_CODES: &str = osc!(104);
+/// Resets the default foreground color.
+pub const RESET_DEFAULT_FG_COLOR: &str = osc!(110);
+/// Resets the default background color.
+pub const RESET_DEFAULT_BG_COLOR: &str = osc!(111);
+/// Resets the cursor color.
+pub const RESET_CURSOR_COLOR: &str = osc!(112);
+
+/// Requests the default foreground color.
+pub const REQUEST_DEFAULT_FG_COLOR: &str = osc!(10, '?');
+/// Requests the default background color.
+pub const REQUEST_DEFAULT_BG_COLOR: &str = osc!(11, '?');
+/// Requests the cursor color.
+pub const REQUEST_CURSOR_COLOR: &str = osc!(12, '?');
+
+/// Requests the copy/paste selection data.
+pub const REQUEST_SELECTION: &str = osc!(52, "", '?');
+
+/// Specifies the selection buffer.
+#[derive(Clone, Debug, Copy, Eq, PartialEq)]
+pub enum Selection {
+    Clipboard,
+    Primary,
+    Secondary,
+    // Either [`Primary`] or [`Clipboard`] (what is the configured default)
+    Select,
+    Cut0,
+    Cut1,
+    Cut2,
+    Cut3,
+    Cut4,
+    Cut5,
+    Cut6,
+    Cut7,
+}
+
+impl Selection {
+    fn get_char(&self) -> char {
+        match self {
+            Selection::Clipboard => 'c',
+            Selection::Primary => 'p',
+            Selection::Secondary => 'q',
+            Selection::Select => 's',
+            Selection::Cut0 => '0',
+            Selection::Cut1 => '1',
+            Selection::Cut2 => '2',
+            Selection::Cut3 => '3',
+            Selection::Cut4 => '4',
+            Selection::Cut5 => '5',
+            Selection::Cut6 => '6',
+            Selection::Cut7 => '7',
+        }
+    }
+}
+
+fn prepare_selection(sel: impl IntoIterator<Item = Selection>) -> String {
+    let mut res = "\x1b]52;".to_string();
+    for b in sel {
+        res.push(b.get_char());
+    }
+    res.push(';');
+    res
+}
+
+/// Requests selection for the first available of the given selection buffers.
+/// If empty requests the default buffer selection.
+pub fn request_selectoin(sel: impl IntoIterator<Item = Selection>) -> String {
+    prepare_selection(sel) + "?\x1b\\"
+}
+
+/// Sets the given selection buffers. If empty sets the default selection
+/// buffers.
+pub fn set_selection(
+    sel: impl IntoIterator<Item = Selection>,
+    data: impl AsRef<[u8]>,
+) -> String {
+    let mut res = prepare_selection(sel);
+    base64::prelude::BASE64_STANDARD.encode_string(data, &mut res);
+    res + "\x1b\\"
+}
+
+// TODO: Kitty extensions
 
 // Internal
 
@@ -493,16 +658,6 @@ pub fn set_cursor(style: CursorStyle) -> &'static str {
 pub const BRACKETED_PASTE_START: &str = "\x1b[200~";
 /// Input code for bracketed paste end. Used internally.
 pub const BRACKETED_PASTE_END: &str = "\x1b[201~";
-
-/*#[macro_export]
-macro_rules! resize_window {
-    ($x:expr, $y:expr) => {
-        $crate::csi!('t', 8, $y, $x)
-    };
-}*/
-
-// TODO: OSC commands
-// TODO: Kitty extensions
 
 /// Trait for getting string from &str and String
 pub trait GetString {